requirements.txt
flask
flask-socketio
openai
soundfile
python-dotenv